chore(tests): update test results in .phpunit.cache/test-results file
The test results in the `.phpunit.cache/test-results` file were updated to include the `TimeSlotsTest` class. This was done to reflect the addition of new tests for the `TimeSlots` resource.

feat(DataFixtures): add loading of default time slots in AppFixtures

The `DefaultTimeSlotsStory` was added to the `AppFixtures` class to load default time slots data. This was done to ensure that the application has some initial time slots data for testing and development purposes.

feat(Factory): add TimeSlotsFactory

The `TimeSlotsFactory` class was added to the `Factory` directory. This factory is responsible for creating instances of the `TimeSlots` entity with default values. It also generates random start and end times for the time slots.

feat(Story): add DefaultTimeSlotsStory

The `DefaultTimeSlotsStory` class was added to the `Story` directory. This story is responsible for creating default time slots using the `TimeSlotsFactory`. It creates 100 time slots with random start and end times.

test(TimeSlotsTest): add tests for TimeSlots resource

The `TimeSlotsTest` class was added to the `tests` directory. It contains tests for the `TimeSlots` resource, including a test for retrieving a collection of time slots and a test for creating a new time slot. These tests ensure that the API endpoints for the `TimeSlots` resource are working correctly.