feat(ansible): add env.j2 template file and update ovh-vm-update.yml playbook
The env.j2 file is added to the ansible directory. This file contains environment variable configurations for the application. It includes variables for APP_ENV, APP_SECRET, SERVER_NAME, and DATABASE_URL.

In the ovh-vm-update.yml playbook, the following changes are made:
- A new task is added to display the value of the APP_ENV environment variable using the ansible.builtin.debug module.
- Another task is added to template the env.j2 file to ~/app/.env with the mode set to "0600". This will create a configuration file for the application with the environment variables.

These changes are made to improve the configuration and deployment process of the application.