feat(ansible): add task to install dependencies using composer in the php container
The task "Install dependencies" has been added to the playbook "ovh-vm-update.yml". This task uses the `community.docker.docker_container_exec` module to execute the command `composer install -o --verbose` inside the `php` container. The purpose of this task is to install the project dependencies with optimized autoloader and display verbose output during the installation process. The result of the task is stored in the `result` variable.